functions/arxiv-llm/requirements.txt
functions-framework==3.*
google-cloud-bigquery>=3.10.0
google-cloud-aiplatform>=1.63.0
vertexai>=1.63.0